refactor common a11y testing pieces
