+++ cypress_test/integration_tests/desktop/writer/a11y_dialog_spec.js
@@ -1,8 +1,8 @@
 /* global describe expect it cy before after afterEach require Cypress */
 
 var helper = require('../../common/helper');
-var ceHelper = require('../../common/contenteditable_helper');
 var desktopHelper = require('../../common/desktop_helper');
+var a11yHelper = require('../../common/a11y_helper');
 
 const allCommonDialogs = [
     '.uno:HyperlinkDialog',
@@ -76,50 +76,17 @@ describe(['tagdesktop'], 'Accessibility Writer Dialog Tests', { testIsolation: f
 
         cy.getFrameWindow().then(function (frameWindow) {
             win = frameWindow;
-
-            const enableUICoverage = {
-                'Track': { 'type': 'boolean', 'value': true }
-            };
-            win.app.map.sendUnoCommand('.uno:UICoverage', enableUICoverage);
+            a11yHelper.enableUICoverage(win);
         });
 
         cy.cGet('.jsdialog-window').should('not.exist');
     });
 
     after(function () {
-        cy.spy(win.app.socket, '_onMessage').as('onMessage').log(false);
+        a11yHelper.reportUICoverage(win);
 
-        // Run after the dialogs are processed and errors checked
-        cy.then(() => {
-            const endUICoverage = {
-                'Report': { 'type': 'boolean', 'value': true },
-                'Track': { 'type': 'boolean', 'value': false }
-            };
-            win.app.map.sendUnoCommand('.uno:UICoverage', endUICoverage);
-        });
-
-        cy.get('@onMessage').should(onMessage => {
-            const matchingCall = onMessage.getCalls().find(call => {
-                const evt = call.args && call.args[0]
-                const textMsg = evt && evt.textMsg;
-                if (!textMsg || !textMsg.startsWith('unocommandresult:')) {
-                    return false;
-                }
-                const jsonPart = textMsg.replace('unocommandresult:', '').trim();
-                const data = JSON.parse(jsonPart);
-                return data.commandName === '.uno:UICoverage';
-            });
-
-            expect(matchingCall, '.uno:UICoverage result').to.be.an('object');
-
-            const textMsg = matchingCall.args[0].textMsg;
-            const jsonPart = textMsg.replace('unocommandresult:', '').trim();
-            const result = JSON.parse(jsonPart).result;
-
-            Cypress.log({name: 'UICoverage Message: ', message: JSON.stringify(result)});
-
+        cy.get('@uicoverageResult').then(result => {
             expect(result.used, `used .ui files`).to.not.be.empty;
-
             // TODO: make these true
             // expect(result.CompleteWriterDialogCoverage, `complete writer dialog coverage`).to.be.true;
             // expect(result.CompleteCommonDialogCoverage, `complete common dialog coverage`).to.be.true;
@@ -138,10 +105,8 @@ describe(['tagdesktop'], 'Accessibility Writer Dialog Tests', { testIsolation: f
             }
         });
         cy.cGet('.jsdialog-window:not(.ui-overflow-group-popup)').should('not.exist');
-        desktopHelper.undoAll();
-        cy.cGet('div.clipboard').as('clipboard');
-        // double click on field at initial cursor position
-        ceHelper.moveCaret('home', 'ctrl');
+
+        a11yHelper.resetState();
     });
 
     // Helper to test that a11y validation detects injected errors
@@ -345,7 +310,7 @@ describe(['tagdesktop'], 'Accessibility Writer Dialog Tests', { testIsolation: f
     }
 
     function runA11yValidation(win) {
-        helper.runA11yValidation(win, 'validatedialogsa11y');
+        a11yHelper.runA11yValidation(win, 'validatedialogsa11y');
     }
 
     function handleDialog(win, level, command) {
